Points animation + color on selected candies + improved point system
Small animation that pop up showing the points that the last match made. Now candies  that are select will have a greenish color just for debbuging (idk why I ignored this for so long). System points are improved, now every candie will have a random value added to the point of the match to make the numbers a little bit more interesting, and the automatic matches (those that happens because the candies that spawn from the top matched alone) will have their score multiplied by 2 for every match,
ex: first match 300 points, the automatic match after that will be 300 x 2, and the other after that will be 300 x 4. The multiplier will be set to 1 after all the automatic matches stop.
Before the automatic matches had the multiplyer set to 2 so:
first matched caused by the player 300 points, first automatic 300 x 2, second automatic 300 x 2 and so on